fix: split handling of `FileNotFoundError` and `subprocess.CalledProcessError`
`subprocess.check_output(...)` can raise two exceptions, `FileNotFoundError` and `subprocess.CalledProcessError`, and they have different public attributes. `pyright` was complaining because the exception handling was being clobbered to handle both errors in the same `except` block and access the `stderr` attribute, but `FileNotFoundError` does not have a `stderr` attribute. This commit splits the error handling into two `except` blocks, and uses a string template to reduce the amount of string duplication for the `ApptainerOpsError` catch-all exception.